auth-dashboard

+++ client/src/Dashboard/DshCategories.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect, useRef, useState } from "react";
 import "./dshmessages.css";
 import DASHHeader from "./DashboardComponents/dashHeader";
 import DashSidebar from "./DashboardComponents/dashSidebar";
@@ -14,14 +14,16 @@ function DshCategories() {
   const [editingCategory, setEditingCategory] = useState(null);
   const [showCreateCity, setShowCreateCity] = useState(false);
   const [newCategory, setNewCategory] = useState({
-    name: "",
+    nameEn: "",
+    nameAr: "",
     imageFile: null,
   });
 
   const [updatedCategory, setUpdatedCategory] = useState({
-    name: "",
+    nameEn: "",
+    nameAr: "",
   });
-
+  const editModalRef = useRef(null);
   const { translations } = useTranslation();
 
   useEffect(() => {
@@ -56,7 +58,8 @@ function DshCategories() {
     try {
       const token = localStorage.getItem("token"); // Retrieve token if stored in localStorage
       const formData = new FormData();
-      formData.append("name", newCategory.name);
+      formData.append("nameEn", newCategory.nameEn);
+      formData.append("nameAr", newCategory.nameAr);
       if (newCategory.imageFile) {
         formData.append("imageFile", newCategory.imageFile);
       }
@@ -68,7 +71,8 @@ function DshCategories() {
       setCategory([...category, response.data]);
 
       setNewCategory({
-        name: "",
+        nameEn: "",
+        nameAr: "",
         imageFile: null,
       });
 
@@ -80,6 +84,8 @@ function DshCategories() {
 
   // Delete category
   const handleDelete = async (id) => {
+    const confirmDelete = window.confirm(`${translations.deleteCat}`);
+    if (!confirmDelete) return;
     try {
       const token = localStorage.getItem("token"); // Get token from storage
       await axios.delete(`${API_BASE_URL}/category/${id}`, {
@@ -97,7 +103,10 @@ function DshCategories() {
   // Open edit modal
   const handleEditClick = (category) => {
     setEditingCategory(category);
-    setUpdatedCategory({ name: category.name });
+    setUpdatedCategory({ nameEn: category.nameEn, nameAr: category.nameAr });
+    setTimeout(() => {
+      editModalRef.current?.scrollIntoView({ behavior: "smooth" });
+    }, 200);
   };
 
   // Update category
@@ -108,7 +117,7 @@ function DshCategories() {
 
       const response = await axios.put(
         `${API_BASE_URL}/category/${editingCategory.id}`,
-        { name: updatedCategory.name }, // Send only the name as JSON
+        { nameEn: updatedCategory.nameEn, nameAr: updatedCategory.nameAr }, // Send only the name as JSON
         {
           headers: {
             Authorization: `Bearer ${token}`,
@@ -153,7 +162,8 @@ function DshCategories() {
               <div className="messageContainer" key={cat.id}>
                 <div className="name">
                   <h3>{cat.id}</h3>
-                  <h3>{cat.name}</h3>
+                  <h3>{cat.nameEn}</h3>
+                  <h3>{cat.nameAr}</h3>
                   {cat.imagePath ? (
                     <img
                       src={"https://" + cat.imagePath}
@@ -186,40 +196,72 @@ function DshCategories() {
                 <h3>Create New category</h3>
                 <input
                   type="text"
-                  placeholder="Category name"
-                  value={newCategory.name}
+                  placeholder="Category name in English"
+                  value={newCategory.nameEn}
                   onChange={(e) =>
                     setNewCategory({
                       ...newCategory,
-                      name: e.target.value,
+                      nameEn: e.target.value,
                     })
                   }
                 />
                 <input
+                  type="text"
+                  placeholder="Category name in Arabic"
+                  value={newCategory.nameAr}
+                  onChange={(e) =>
+                    setNewCategory({
+                      ...newCategory,
+                      nameAr: e.target.value,
+                    })
+                  }
+                />
+                <input
+                  className="file"
                   type="file"
                   onChange={(e) => handleFileChange(e, "new")}
                 />
-                <button onClick={handleCreateCategory}>Create Category</button>
+                <button className="addprod" onClick={handleCreateCategory}>
+                  Create Category
+                </button>
               </div>
             )}
 
             {/* Edit category Modal */}
             {editingCategory && (
-              <div className="edit-category-modal">
+              <div ref={editModalRef} className="edit-category-modal">
                 <h3>Edit category</h3>
                 <input
                   type="text"
-                  placeholder="Category name"
-                  value={updatedCategory.name}
+                  placeholder="Category name in English"
+                  value={updatedCategory.nameEn}
                   onChange={(e) =>
                     setUpdatedCategory({
                       ...updatedCategory,
-                      name: e.target.value,
+                      nameEn: e.target.value,
                     })
                   }
                 />
-                <button onClick={handleUpdate}>Update</button>
-                <button onClick={() => setEditingCategory(null)}>Cancel</button>
+                <input
+                  type="text"
+                  placeholder="Category name in Arabic"
+                  value={updatedCategory.nameAr}
+                  onChange={(e) =>
+                    setUpdatedCategory({
+                      ...updatedCategory,
+                      nameAr: e.target.value,
+                    })
+                  }
+                />
+                <button className="addprod" onClick={handleUpdate}>
+                  Update
+                </button>
+                <button
+                  className="addprod"
+                  onClick={() => setEditingCategory(null)}
+                >
+                  Cancel
+                </button>
               </div>
             )}
           </main>
